Fix maven property version
Since the project uses the spring boot starter already, there is no
reason to create another version for Neo4j. The one in Spring Boot is
`neo4j-ogm.version` and it can be reused to add additional artifacts.

The added bonus is that such override is going to be transparently
handled if the version of Spring Boot changes. Right now, it would
require that version to be kept in sync with the Spring Boot version.